+++ src/main/java/com/example/high_five/dto/point/PointAdminPolicyRequest.java
@@ -3,8 +3,10 @@ package com.example.high_five.dto.point;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
+@Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class PointAdminPolicyRequest {
