+++ src/main/java/com/example/high_five/service/CouponService.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.Map;
 
-@FeignClient(name = "gateway-server", contextId = "couponClient", url = "http://gateway-server:8000")
+@FeignClient(name = "gateway-server", contextId = "couponClient", url = "http://localhost:8082")
 public interface CouponService {
 
     @GetMapping("/api/coupons/members/{memberId}")
