+++ src/main/java/com/example/high_five/service/MemberService.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 
-@FeignClient(name = "gateway-server", contextId = "memberClient", url = "http://gateway-server:8000")
+@FeignClient(name = "gateway-server", contextId = "memberClient", url = "http://localhost:8082")
 public interface MemberService {
     // 관리자 - 정책 조회
     @GetMapping("/api/admin/points/policy")
