+++ src/main/resources/templates/admin/points.html
@@ -187,19 +187,20 @@
             <div th:if="${message}" class="alert-msg alert-success" th:text="${message}"></div>
             <div th:if="${errorMessage}" class="alert-msg alert-error" th:text="${errorMessage}"></div>
 
-            <form method="post" action="/api/admin/points/policy">
-
+            <form method="post" th:action="@{/api/admin/points/policy}">
                 <div class="form-row">
                     <label>회원가입 적립 P</label>
-                    <input type="number" name="signupPoint" th:value="${policy.signupPoint ?: 0}">
+                        <input type="number" name="signupPoint" th:value="${policy.signupPoint ?: 0}">
                 </div>
+
                 <div class="form-row">
                     <label>일반 리뷰 적립 P</label>
-                    <input type="number" name="reviewPoint" th:value="${policy.reviewPoint ?: 0}">
+                        <input type="number" name="reviewPoint" th:value="${policy.reviewPoint ?: 0}">
                 </div>
+
                 <div class="form-row">
                     <label>포토 리뷰 적립 P</label>
-                    <input type="number" name="photoPoint" th:value="${policy.photoPoint ?: 0}">
+                        <input type="number" name="photoPoint" th:value="${policy.photoPoint ?: 0}">
                 </div>
 
                 <input type="submit" value="정책 저장하기" class="save-btn">
@@ -209,21 +210,25 @@
 
         <div class="card">
             <h3>포인트 수동 조정</h3>
-            <form method="post" action="/api/admin/points/adjustment">
+            <form method="post" th:action="@{/api/admin/points/adjustment}">
                 <div class="form-row">
                     <label>회원 ID</label>
-                    <input type="number" name="memberId" placeholder="회원번호 입력">
+                        <input type="number" name="memberId" placeholder="회원번호 입력">
                 </div>
+
                 <div class="form-row">
                     <label>조정 금액 (+/-)</label>
-                    <input type="number" name="amount" placeholder="예: 1000 or -500">
+                        <input type="number" name="amount" placeholder="예: 1000 or -500">
                 </div>
+
                 <div class="form-row">
                     <label>사유</label>
-                    <input type="text" name="reason" placeholder="예: 배송 지연 보상" style="text-align: left;">
+                        <input type="text" name="reason" placeholder="예: 배송 지연 보상" style="text-align: left;">
                 </div>
+
                 <input type="submit" value="포인트 조정 실행" class="save-btn" style="background-color: #e9c46a;">
             </form>
+
         </div>
 
     </main>
